break: make themes independent due to overwrite css custom properties + move themes in individual directories each
