+++ packages/dnb-ui-lib/scripts/prepub/tasks/__tests__/themeFactory.test.js
@@ -5,24 +5,24 @@
 
 import path from 'path'
 import { runFactory } from '../themeFactory'
-import '../../../../src/style/themes/dnb-theme-ui.scss' // just to make sure we re-run the test in watch mode due to changes in this file
+import '../../../../src/style/themes/theme-ui/dnb-theme-ui.scss' // just to make sure we re-run the test in watch mode due to changes in this file
 
 beforeAll(async () => {
   global.css = (await runFactory({
     processToNamesList: [
       path.resolve(
         __dirname,
-        '../../../../src/style/themes/dnb-theme-ui.scss'
+        '../../../../src/style/themes/theme-ui/dnb-theme-ui.scss'
       )
     ],
     returnResult: true
   }))[0]
 })
 
-describe('themeFactory collect and generate the themes', () => {
+describe('The main UI Theme', () => {
   it('has to have the default theme with correct path', () => {
     expect(global.css).toContain(
-      `@import '../../style/themes/dnb-theme-ui.scss';`
+      `@import '../../../style/themes/theme-ui/dnb-theme-ui.scss';`
     )
   })
 })
