--- packages/dnb-ui-lib/src/style/themes/dnb-theme-ui.scss
@@ -1,50 +1,0 @@
-/**
- * Main UI Theme
- *
- */
-
-$fonts-path: '../../../assets/fonts/' !default;
-
-@import '../core/utilities.scss';
-
-// define the default ui theme
-@import './theme-ui/anchor.scss';
-@import './theme-ui/typography.scss';
-@import './theme-ui/blockquote.scss';
-@import './theme-ui/lists.scss';
-@import './theme-ui/table.scss';
-@import './theme-ui/label.scss';
-@import './theme-ui/textarea.scss';
-@import './theme-ui/form.scss';
-@import './theme-ui/code.scss';
-@import './theme-ui/image.scss';
-
-// @import './dnb-theme-ui-basis.scss';
-@import './theme-ui/ui-elements.scss';
-$preventIncludeElements: false !default;
-@if $preventIncludeElements != true {
-  @include includeElements();
-}
-
-// make sure we load the spacings after we have defined the defaults of the elements
-@import './theme-ui/ui-spacing.scss';
-
-@supports not (-ms-ime-align: auto) {
-  ::selection {
-    background-color: var(--color-mint-green);
-    color: var(--color-black);
-    text-shadow: none;
-  }
-}
-
-/**
-  ATTENTION: This file is auto generated by using "styleFactory".
-  Do not change the content above!
-  All the themes get auto generated in here
-*/
-
-@import '../../components/button/style/themes/dnb-button-theme-ui.scss';
-@import '../../components/form-status/style/themes/dnb-form-status-theme-ui.scss';
-@import '../../components/input/style/themes/dnb-input-theme-ui.scss';
-@import '../../components/step-indicator/style/themes/dnb-step-indicator-theme-ui.scss';
-@import '../../components/tabs/style/themes/dnb-tabs-theme-ui.scss';
