+++ packages/dnb-ui-lib/src/style/themes/theme-open-banking/dnb-theme-open-banking.scss
@@ -4,18 +4,20 @@
  *
  */
 
-// import the default theme, and go from there
-// will also have all the component themes included
-@import './dnb-theme-ui.scss';
-
-:root {
+// change some properties
+@import '../../core/properties.scss';
+@include setProperties() {
   // open-bankin colors
   --color-open-banking-green: #a5e1d2;
   --color-open-banking-emerald: #00343e;
 
   // Overwriting UX colors, defined in properties.scss
-  --color-sea-green: fuchsia;
+  --color-sea-green: fuchsia; //#f0f
 }
+
+// import the default theme, and go from there
+// will also have all the component themes included
+@import '../theme-ui/dnb-theme-ui.scss';
 
 /**
   ATTENTION: This file is auto generated by using "themeFactory".
@@ -23,4 +25,4 @@
   All the themes get auto generated in here
 */
 
-@import '../../components/button/style/themes/dnb-button-theme-open-banking.scss';
+@import '../../../components/button/style/themes/dnb-button-theme-open-banking.scss';
