+++ packages/dnb-ui-lib/src/style/themes/theme-ui/dnb-theme-ui.scss
@@ -1,0 +1,51 @@
+/**
+ * Main UI Theme
+ *
+ */
+
+$fonts-path: '../../../../assets/fonts/' !default;
+
+@import '../../core/utilities.scss';
+@import '../../core/properties.scss';
+
+// define the default ui theme
+@import './anchor.scss';
+@import './typography.scss';
+@import './blockquote.scss';
+@import './lists.scss';
+@import './table.scss';
+@import './label.scss';
+@import './textarea.scss';
+@import './form.scss';
+@import './code.scss';
+@import './image.scss';
+
+// @import './dnb-theme-ui-basis.scss';
+@import './ui-elements.scss';
+$preventIncludeElements: false !default;
+@if $preventIncludeElements != true {
+  @include includeElements();
+}
+
+// make sure we load the spacings after we have defined the defaults of the elements
+@import './ui-spacing.scss';
+
+@supports not (-ms-ime-align: auto) {
+  ::selection {
+    background-color: var(--color-mint-green);
+    color: var(--color-black);
+    text-shadow: none;
+  }
+}
+
+/**
+  ATTENTION: This file is auto generated by using "styleFactory".
+  Do not change the content above!
+  All the themes get auto generated in here
+*/
+
+@import '../../../components/button/style/themes/dnb-button-theme-ui.scss';
+@import '../../../components/form-status/style/themes/dnb-form-status-theme-ui.scss';
+@import '../../../components/input/style/themes/dnb-input-theme-ui.scss';
+@import '../../../components/step-indicator/style/themes/dnb-step-indicator-theme-ui.scss';
+@import '../../../components/tabs/style/themes/dnb-tabs-theme-ui.scss';
